Split up the SafetyExecutionLogic object into smaller traits
