+++ src/main/scala/com/leobenkel/safetyplugin/Transformations/TaskConfiguration.scala
@@ -1,0 +1,19 @@
+package com.leobenkel.safetyplugin.Transformations
+
+import com.leobenkel.safetyplugin.{Config, SafetyPluginKeys}
+import com.leobenkel.safetyplugin.SafetyPluginKeys.safetyConfPath
+import sbt.Def
+
+private[Transformations] trait TaskConfiguration {
+  def safetyConfigurationExec(): Def.Initialize[Config.SafetyConfiguration] = {
+    Def.settingDyn {
+      val log = SafetyPluginKeys.safetyGetLog.value
+      val path = safetyConfPath.value
+
+      Def.setting {
+        val conf = Config.ConfigurationParser(log, path)
+        conf.getConf
+      }
+    }
+  }
+}
